db: temporary place db impl here
Since the project has not even been creaed, the implementations
are temporary copied to here; when the project is created, add
them into the project.

The SQL scripts are removed since Android Room is an ORM capable
of doing those work. Pre-populated data will need to be re-written
because of the change of schema.